+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Enhancements
     - Usage of `Supervisor` functions becoming deprecated in Elixir `1.11.0` has been removed
 
+- Backwards incompatible changes
+    - `:mmdb2_decoder` has been updated to `~> 3.0`
+
 ## v0.4.0 (2020-04-12)
 
 - Enhancements